Build and launch on hardware; kernel still hangs on device
The extension now compiles, links, and launches on a real 910B3. It does not
yet complete: a single tile (B=1 T=16 H=1) hangs deterministically. Recorded in
docs/debugging-notes.md rather than left to be rediscovered.

Build fixes:
- layout.hpp is shared by device kernels and the host launch TU, so its
  AscendC/catlass includes are gated on __CCE__ with host stand-ins for
  GM_ADDR and the element types.
- Workspace sizing helpers move from the device-only utils.hpp into
  layout.hpp beside WorkspaceSizes, which also keeps the host allocation and
  the kernel offsets from drifting.
- Host casts drop the __gm__ address-space qualifier (FwdParams carries plain
  byte pointers) and use is_privateuseone(), since torch_npu exposes NPU
  tensors through the PrivateUse1 backend.
- The pybind binding needs torch/extension.h for at::Tensor casters and
  pybind11/stl.h for std::optional; without both every tensor argument is
  rejected at the boundary.

beta is no longer transposed host-side. It is [T_total, H] natively and a
chunk's 16 values sit at stride H, which DataCopyPad handles directly. The
transpose was materializing a device copy on every call.

Test harness: this image ships the AscendC compiler but not the precompiled
built-in operator library, so every torch_npu compute op fails to launch --
even torch.randn on device. Allocation and H2D/D2H copies work, and custom
AscendC kernels work (catlass 00_basic_matmul reports Compare success). So
tests/test_npu_nocompute.py generates inputs and computes the reference on CPU
and uses the device only to hold inputs and return results. torch_ref.py loses
one NPU-only torch.mm(out_dtype=) that is a no-op for fp32 inputs, and the
tests now import torch_npu so the npu device is registered at all.

tests/sync_probe.cpp is a standalone probe that runs kernel1's exact two-round
AIC/AIV handshake with the compute stripped out. It completes, which rules the
handshake pattern out as the cause.

+++ FlashKDA-Ascend/tests/torch_ref.py
@@ -10,6 +10,7 @@ No CUDA dependencies — works on CPU or NPU.
 """
 
 import torch
+import torch_npu  # registers the npu device with torch
 import math
 
 # ============================================================
@@ -224,7 +225,7 @@ def torch_ref(q, k, v, g, beta, scale, out, A_log, dt_bias, lower_bound,
                 _out = _out + torch.matmul(Mqk, U)
 
                 # State update: state = state * exp(g_total) + k_restored^T @ u
-                delta_s = torch.mm(k_restored.t(), U, out_dtype=torch.float32)
+                delta_s = torch.mm(k_restored.t().float(), U.float())
 
                 g_total_exp = fp32_ex2_ftz(g_total)
                 g_total_exp = g_total_exp.squeeze(0).unsqueeze(-1)
